New ReservationItem component
